Migrated rsi indicator to the class pattern.
