+++ js/masters/indicators/indicators-all.src.js
@@ -40,7 +40,7 @@ import '../../Stock/Indicators/PCIndicator.js';
 import '../../Stock/Indicators/PriceEnvelopesIndicator.js';
 import '../../Stock/Indicators/PSARIndicator.js';
 import '../../Stock/Indicators/ROCIndicator.js';
-import '../../Stock/Indicators/RSIIndicator.js';
+import '../../Stock/Indicators/RSI/RSIIndicator.js';
 import '../../Stock/Indicators/StochasticIndicator.js';
 import '../../Stock/Indicators/SlowStochasticIndicator.js';
 import '../../Stock/Indicators/SupertrendIndicator.js';
