+++ ts/Stock/Indicators/RSI/RSIPoint.d.ts
@@ -1,0 +1,34 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type RSIIndicator from './RSIIndicator';
+import type SMAPoint from '../SMA/SMAPoint';
+
+/* *
+ *
+ *  Class
+ *
+ * */
+
+declare class RSIPoint extends SMAPoint {
+    public series: RSIIndicator;
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default RSIPoint;
